Renamed some classes and variables
