+++ client/src/main/scala/it/cwmp/client/view/FXInputView.scala
@@ -3,7 +3,7 @@ package it.cwmp.client.view
 /**
   * A trait that gives generic methods that all JavaFX input controllers should have
   */
-trait FXInputViewController extends FXViewController {
+trait FXInputView {
 
   /**
     * Resets input fields
