+++ client/src/main/scala/it/cwmp/client/view/room/RoomFXController.scala
@@ -23,8 +23,8 @@ object RoomFXController {
   }
 }
 
-class RoomFXController(strategy: RoomFXStrategy) extends FXInputViewController
-  with FXView with FXInputChecks with FXAlerts with FXRunOnUIThread {
+class RoomFXController(strategy: RoomFXStrategy) extends FXInputView
+  with FXViewController with FXInputChecks with FXAlerts with FXRunOnUIThread {
 
   protected val layout: String = LayoutRes.roomManagerLayout
   protected val title: String = StringRes.roomManagerTitle
